scGen and scVIDR fix
