+++ configs/models/scgen.yaml
@@ -1,4 +1,4 @@
-name: scVIDR
+name: vae_scgen
 model:
   latent_dim: 100
   hidden_dim: 800
@@ -6,7 +6,6 @@ model:
 training:
   dropout_rate: 0.2
   learning_rate: 0.001
-  alpha: 0.001
-  epochs: 50
-  delta_predictor_epochs: 300
-  batsize: 128
+  alpha: 0.00005
+  epochs: 100
+  batsize: 32
