--- configs/models/vae_scgen.yaml
@@ -1,11 +1,0 @@
-name: vae_scgen
-model:
-  latent_dim: 20
-  hidden_dim: 800
-
-training:
-  dropout_rate: 0.75
-  learning_rate: 0.1
-  alpha: 0.01
-  epochs: 100
-  batsize: 256
